Refactor TransmissionParameters
* Moves all TransmissionParameter types into `connections` to avoid some
  circular imports.
* Splits up `EnsembleTransmissionParameters` so that transforms and the
  original decoders are remembered.
