+++ nengo_spinnaker/builder/connection.py
@@ -1,4 +1,6 @@
 import nengo
+import numpy as np
+
 from .builder import Model, ObjectPort, spec
 from .model import ReceptionParameters, InputPort, OutputPort
 
@@ -23,3 +25,95 @@ def build_generic_reception_params(model, conn):
     """
     # Just extract the synapse from the connection.
     return ReceptionParameters(conn.synapse, conn.post_obj.size_in)
+
+
+class EnsembleTransmissionParameters(object):
+    """Transmission parameters for a connection originating at an Ensemble.
+
+    Attributes
+    ----------
+    decoders : array
+        Decoders to use for the connection.
+    """
+    def __init__(self, decoders, transform):
+        # Copy the decoders
+        self.untransformed_decoders = np.array(decoders)
+        self.transform = np.array(transform)
+
+        # Compute and store the transformed decoders
+        self.decoders = np.dot(transform, decoders.T).T
+
+        # Make the arrays read-only
+        self.untransformed_decoders.flags['WRITEABLE'] = False
+        self.transform.flags['WRITEABLE'] = False
+        self.decoders.flags['WRITEABLE'] = False
+
+    def __ne__(self, other):
+        return not (self == other)
+
+    def __eq__(self, other):
+        # Equal iff. the objects are of the same type
+        if type(self) is not type(other):
+            return False
+
+        # Equal iff. the decoders are the same shape
+        if self.decoders.shape != other.decoders.shape:
+            return False
+
+        # Equal iff. the decoder values are the same
+        if np.any(self.decoders != other.decoders):
+            return False
+
+        return True
+
+
+class PassthroughNodeTransmissionParameters(object):
+    """Parameters describing connections which originate from pass through
+    Nodes.
+    """
+    def __init__(self, transform):
+        # Store the parameters, copying the transform
+        self.transform = np.array(transform)
+
+    def __ne__(self, other):
+        return not (self == other)
+
+    def __eq__(self, other):
+        # Equivalent if the same type
+        if type(self) is not type(other):
+            return False
+
+        # and the transforms are equivalent
+        if (self.transform.shape != other.transform.shape or
+                np.any(self.transform != other.transform)):
+            return False
+
+        return True
+
+
+class NodeTransmissionParameters(PassthroughNodeTransmissionParameters):
+    """Parameters describing connections which originate from Nodes."""
+    def __init__(self, pre_slice, function, transform):
+        # Store the parameters
+        super(NodeTransmissionParameters, self).__init__(transform)
+        self.pre_slice = pre_slice
+        self.function = function
+
+    def __hash__(self):
+        # Hash by ID
+        return hash(id(self))
+
+    def __eq__(self, other):
+        # Parent equivalence
+        if not super(NodeTransmissionParameters, self).__eq__(other):
+            return False
+
+        # Equivalent if the pre_slices are exactly the same
+        if self.pre_slice != other.pre_slice:
+            return False
+
+        # Equivalent if the functions are the same
+        if self.function is not other.function:
+            return False
+
+        return True
